+++ Changelog.md
@@ -1,7 +1,19 @@
+### Changelogs:
+
+#### v0.0.3
+##### Updates
+* Minor updates
+* Upload and Download added - needs maintain
+* Post method fixed and added back
+
+#### v0.0.2
+##### Updates
+* Minor updates
+* Test files added
+* Removed `Post` method temporarily
+
 #### v0.0.1
-
 ##### Publish
-
 * Available get Method
 * Incoming {Upload|Download} files